+++ news/news.rst
@@ -21,4 +21,19 @@
   to tell the user to finish the interactive calibration process in the
   analysis terminal.
 
+* Fix ``_end_beamtime``. Details about the fixes are:
+
+  * Use rsync while archiving ``xpdUser`` so that user can see 
+    the progress. (rsync lists files have been transferred)
+
+  * More sophisticated logic when flushing xpdUser directory. 
+    Now the function will tell the user to close files used by 
+    the current process, instead of throwing an error and failing 
+    the process.
+
+  * Some cleaning in the logic. Program will remove the remote 
+    archive if user doesn't confirm to flush the local directory 
+    so that we could potentially avoid having multiple copies at 
+    the remote location.
+
 **Security:** None